Rename Reporter -> AnalyticsReporter
