+++ app/src/main/java/org/simple/clinic/ReleaseClinicApp.kt
@@ -2,7 +2,7 @@ package org.simple.clinic
 
 import android.annotation.SuppressLint
 import org.simple.clinic.analytics.Analytics
-import org.simple.clinic.analytics.HeapReporter
+import org.simple.clinic.analytics.HeapAnalyticsReporter
 import org.simple.clinic.di.AppComponent
 import org.simple.clinic.di.AppModule
 import org.simple.clinic.di.DaggerAppComponent
@@ -18,7 +18,7 @@ class ReleaseClinicApp : ClinicApp() {
   override fun onCreate() {
     super.onCreate()
     appComponent.inject(this)
-    Analytics.addReporter(HeapReporter(this))
+    Analytics.addReporter(HeapAnalyticsReporter(this))
     keepUserIdUpdatedInAnalytics()
     syncScheduler.schedule().subscribe()
   }
