+++ app/src/main/java/org/simple/clinic/analytics/DebugAnalyticsReporter.kt
@@ -2,7 +2,7 @@ package org.simple.clinic.analytics
 
 import timber.log.Timber
 
-class DebugReporter : Reporter {
+class DebugAnalyticsReporter : AnalyticsReporter {
 
   override fun setUserIdentity(id: String) {
     Timber.tag("Analytics").d("User ID: $id")
