+++ app/src/test/java/org/simple/clinic/summary/PatientSummaryScreenControllerTest.kt
@@ -19,7 +19,7 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.simple.clinic.analytics.Analytics
-import org.simple.clinic.analytics.MockReporter
+import org.simple.clinic.analytics.MockAnalyticsReporter
 import org.simple.clinic.bp.BloodPressureMeasurement
 import org.simple.clinic.bp.BloodPressureRepository
 import org.simple.clinic.drugs.PrescriptionRepository
@@ -54,7 +54,7 @@ class PatientSummaryScreenControllerTest {
   private val clock = Clock.fixed(Instant.now(), UTC)
 
   private val uiEvents = PublishSubject.create<UiEvent>()
-  private val reporter = MockReporter()
+  private val reporter = MockAnalyticsReporter()
 
   /**
    * If the placeholder count is changed, the `params for placeholder bp items`
@@ -340,7 +340,7 @@ class PatientSummaryScreenControllerTest {
   fun `when the screen is opened, the viewed patient analytics event must be sent`(fromCaller: PatientSummaryCaller) {
     uiEvents.onNext(PatientSummaryScreenCreated(patientUuid, fromCaller))
 
-    val expectedEvent = MockReporter.Event("ViewedPatient", mapOf(
+    val expectedEvent = MockAnalyticsReporter.Event("ViewedPatient", mapOf(
         "patientId" to patientUuid.toString(),
         "from" to fromCaller.name
     ))
